requirements.txt
azure-cognitiveservices-speech==1.40.0
azure-ai-textanalytics==5.3.0
python-dotenv==1.0.0
flask==3.0.0
